+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ConditionLayoutCategory` component. 
 
 ## [2.4.0] - 2021-12-28
 
@@ -19,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [2.2.0] - 2021-07-28
 ### Added
 - `ConditionLayoutBinding` component. 
-
+ 
 ## [2.1.3] - 2021-01-27
 ### Fixed
 - Avoid breaking if product context is not filled yet.